seg_requirements.txt
torch
torchvision
pandas
numpy
Pillow
tqdm
scikit-learn
pycm
matplotlib
timm==0.3.2
